+++ src/api/layers/controls.layer.ts
@@ -63,6 +63,26 @@ export class ControlsLayer extends UILayer {
   constructor(public page: Page, session?: string, options?: CreateConfig) {
     super(page, session, options);
   }
+  /**
+   * Check if there is chat
+   * @param contactId {string} id '000000000000@c.us'
+   * @returns object
+   */
+  public async checkChat(contactId: string) {
+    return new Promise(async (resolve, reject) => {
+      const result = await this.page.evaluate(
+        ({ contactId }) => {
+          return WAPI.checkChat(contactId);
+        },
+        { contactId }
+      );
+      if (result['erro'] == true) {
+        return reject(result);
+      } else {
+        return resolve(result);
+      }
+    });
+  }
 
   /**
    * Unblock contact
@@ -161,17 +181,17 @@ export class ControlsLayer extends UILayer {
   }
 
   /**
-   * Deletes message of given message id
+   * Deletes message all of given message id
    * @param chatId The chat id from which to delete the message.
    * @param messageId The specific message id of the message to be deleted
    * @param onlyLocal If it should only delete locally (message remains on the other recipienct's phone). Defaults to false.
    */
-  public async deleteMessage(
+  public async deleteMessageAll(
     chatId: string,
-    messageId: string[]
+    messageId: string[] | string
   ): Promise<Object> {
     return new Promise(async (resolve, reject) => {
-      const typeFunction = 'deleteMessage';
+      const typeFunction = 'deleteMessageAll';
       const type = 'string';
       const check = [
         {
@@ -195,8 +215,57 @@ export class ControlsLayer extends UILayer {
         return reject(validating);
       }
       const result = await this.page.evaluate(
-        ({ chatId, messageId }) => WAPI.deleteMessages(chatId, messageId),
-        { chatId, messageId }
+        ({ contactId, messageId }) =>
+          WAPI.deleteMessagesAll(contactId, messageId, false),
+        { contactId: chatId, messageId }
+      );
+
+      if (result['erro'] == true) {
+        return reject(result);
+      } else {
+        return resolve(result);
+      }
+    });
+  }
+
+  /**
+   * Deletes message me of given message id
+   * @param chatId The chat id from which to delete the message.
+   * @param messageId The specific message id of the message to be deleted
+   * @param onlyLocal If it should only delete locally (message remains on the other recipienct's phone). Defaults to false.
+   */
+  public async deleteMessageMe(
+    chatId: string,
+    messageId: string[] | string
+  ): Promise<Object> {
+    return new Promise(async (resolve, reject) => {
+      const typeFunction = 'deleteMessageMe';
+      const type = 'string';
+      const check = [
+        {
+          param: 'chatId',
+          type: type,
+          value: chatId,
+          function: typeFunction,
+          isUser: true
+        },
+        {
+          param: 'messageId',
+          type: 'object',
+          value: messageId,
+          function: typeFunction,
+          isUser: true
+        }
+      ];
+
+      const validating = checkValuesSender(check);
+      if (typeof validating === 'object') {
+        return reject(validating);
+      }
+      const result = await this.page.evaluate(
+        ({ contactId, messageId }) =>
+          WAPI.deleteMessagesMe(contactId, messageId, true),
+        { contactId: chatId, messageId }
       );
 
       if (result['erro'] == true) {