+++ src/lib/wapi/functions/check-send-exist.js
@@ -52,36 +52,43 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNMMNMNMMMNMMNNMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
-export function scope(id, erro, status, text = null) {
-  let e = {
-    me: Store.Me.attributes,
+export function scope(id, erro, status, text = null, result = null) {
+  const object = {
+    me: WAPI.getHost(),
     to: id,
     erro: erro,
     text: text,
-    status: status
+    status: status,
+    result: result
   };
-  return e;
+  return object;
 }
 export async function getchatId(chatId) {
-  var to = await WAPI.getChatById(chatId);
-  if (to) {
-    var objTo = to.lastReceivedKey,
-      extend = {
-        formattedName: to.contact.formattedName,
-        isBusiness: to.contact.isBusiness,
-        isMyContact: to.contact.isMyContact,
-        verifiedName: to.contact.verifiedName,
-        pushname: to.contact.pushname,
-        isOnline: to.isOnline
-      };
-    Object.assign(objTo, extend);
-    return objTo;
-  } else {
-    return undefined;
+  if (chatId) {
+    let to = await WAPI.getChatById(chatId);
+    if (to && typeof to === 'object') {
+      let objTo = to.lastReceivedKey;
+      if (objTo && typeof objTo === 'object') {
+        let extend = {
+          formattedName: to.contact.formattedName,
+          isBusiness: to.contact.isBusiness,
+          isMyContact: to.contact.isMyContact,
+          verifiedName: to.contact.verifiedName,
+          pushname: to.contact.pushname,
+          isOnline: to.isOnline
+        };
+        Object.assign(objTo, extend);
+        return objTo;
+      }
+    }
   }
+  return undefined;
 }
 
-export async function sendExist(chatId, returnChat = true, Send = true) {
+export function sendCheckType(chatId = undefined) {
+  if (!chatId) {
+    return WAPI.scope(chatId, true, 404, 'It is necessary to pass a number!');
+  }
   if (typeof chatId === 'string') {
     const contact = '@c.us';
     const broadcast = '@broadcast';
@@ -127,24 +134,38 @@ export async function sendExist(chatId, returnChat = true, Send = true) {
 
     if (
       grup === chatId.substr(-grup.length, grup.length) &&
-      ((chatId.match(/(@g.us)/g) && chatId.match(/(@g.us)/g).length > 1) ||
-        !chatId.match(/^\d+(-)+(\d)*@g.us$/g))
+      chatId.match(/(@g.us)/g) &&
+      chatId.match(/(@g.us)/g).length > 1
     ) {
       return WAPI.scope(
         chatId,
         true,
         404,
-        'incorrect parameters! Use as an example: 00000000-000000@g.us'
+        'incorrect parameters! Use as an example: 00000000-000000@g.us or 00000000000000@g.us'
       );
     }
   }
+}
 
-  let ck = await window.WAPI.checkNumberStatus(chatId);
+export async function sendExist(chatId, returnChat = true, Send = true) {
+  const checkType = WAPI.sendCheckType(chatId);
+  if (!!checkType && checkType.status === 404) {
+    return checkType;
+  }
+  let ck = await window.WAPI.checkNumberStatus(chatId, false);
 
-  if (ck.status === 404) {
+  if (
+    ck.status === 404 &&
+    !chatId.includes('@g.us') &&
+    !chatId.includes('@broadcast')
+  ) {
     return WAPI.scope(chatId, true, ck.status, 'The number does not exist');
   }
-  let chat = await window.WAPI.getChat(ck.id._serialized);
+
+  let chat =
+    ck && ck.id && ck.id._serialized
+      ? await window.WAPI.getChat(ck.id._serialized)
+      : undefined;
 
   if (ck.numberExists && chat === undefined) {
     var idUser = new window.Store.UserConstructor(chatId, {
@@ -156,7 +177,10 @@ export async function sendExist(chatId, returnChat = true, Send = true) {
   if (!chat) {
     const storeChat = await window.Store.Chat.find(chatId);
     if (storeChat) {
-      chat = await window.WAPI.getChat(storeChat.id._serialized);
+      chat =
+        storeChat && storeChat.id && storeChat.id._serialized
+          ? await window.WAPI.getChat(storeChat.id._serialized)
+          : undefined;
     }
   }
 
@@ -173,7 +197,13 @@ export async function sendExist(chatId, returnChat = true, Send = true) {
     );
   }
 
-  if (!ck.numberExists && !chat.t && chat.isBroadcast) {
+  if (
+    !ck.numberExists &&
+    !chat.t &&
+    chat.id &&
+    chat.id.user != 'status' &&
+    chat.isBroadcast
+  ) {
     return WAPI.scope(
       chatId,
       true,
@@ -183,7 +213,7 @@ export async function sendExist(chatId, returnChat = true, Send = true) {
   }
 
   if (!chat) {
-    return WAPI.scope(ck.id._serialized, true, 404);
+    return WAPI.scope(chatId, true, 404);
   }
 
   if (Send) {
@@ -194,5 +224,5 @@ export async function sendExist(chatId, returnChat = true, Send = true) {
     return chat;
   }
 
-  return WAPI.scope(ck.id._serialized, false, 200);
+  return WAPI.scope(chatId, false, 200);
 }