+++ src/lib/wapi/functions/reply.js
@@ -107,18 +107,21 @@ export async function reply(chatId, content, quotedMessageId) {
       quotedMsgOptions = quotedMessage.msgContextInfo(chat);
     }
 
-    let checkID = await WAPI.checkIdMessage(chat.id, quotedMessageId);
+    let checkID = await WAPI.checkIdMessage(
+      chat.id._serialized,
+      quotedMessageId
+    );
     if (checkID.erro == true) {
       return checkID;
     }
 
-    const newMsgId = await window.WAPI.getNewMessageId(chat.id);
+    const newMsgId = await window.WAPI.getNewMessageId(chat.id._serialized);
     let inChat = await WAPI.getchatId(chat.id).catch(() => {});
     if (inChat) {
       chat.lastReceivedKey._serialized = inChat._serialized;
       chat.lastReceivedKey.id = inChat.id;
     }
-    const fromwWid = await window.Store.Conn.wid;
+    const fromwWid = await Store.MaybeMeUser.getMaybeMeUser();
 
     const message = {
       id: newMsgId,