+++ src/lib/wapi/store/store-objects.js
@@ -209,7 +209,7 @@ export const storeObjects = [
   {
     id: 'Me',
     conditions: (module) =>
-      module.PLATFORMS && module.Conn ? module.default : null
+      module.PLATFORMS && module.Conn ? module.Conn : null,
   },
   {
     id: 'CallUtils',
@@ -227,13 +227,18 @@ export const storeObjects = [
       module.getStatus && module.setMyStatus ? module : null
   },
   {
-    id: 'ChatStates',
+    id: 'ChatState',
     conditions: (module) =>
       module.sendChatStatePaused &&
       module.sendChatStateRecording &&
       module.sendChatStateComposing
         ? module
         : null
+  },
+  {
+    id: 'sendDeleteMsgs',
+    conditions: (module) =>
+      module.sendDeleteMsgs ? module.sendDeleteMsgs : null
   },
   {
     id: 'GroupActions',
@@ -293,16 +298,6 @@ export const storeObjects = [
       module.parseMsgStubProto &&
       module.binSend &&
       module.subscribeLiveLocation
-        ? module
-        : null
-  },
-  {
-    id: 'Versions',
-    conditions: (module) =>
-      module.loadProtoVersions &&
-      module.default['15'] &&
-      module.default['16'] &&
-      module.default['17']
         ? module
         : null
   },
@@ -406,6 +401,10 @@ export const storeObjects = [
       module.default && module.default.destroyStorage ? module.default : null
   },
   {
+    id: 'Login',
+    conditions: (module) => (module.startLogout ? module : null),
+  },
+  {
     id: 'BlobCache',
     conditions: (module) =>
       module.default && module.default.getOrCreateURL ? module.default : null
@@ -425,5 +424,43 @@ export const storeObjects = [
   {
     id: 'GroupSettings',
     conditions: (module) => (module.sendSetGroupProperty ? module : null)
-  }
+  },
+  {
+    id: 'MaybeMeUser',
+    conditions: (module) => (module.getMaybeMeUser ? module : null)
+  },
+  {
+    id: 'sendCreateGroup',
+    conditions: (module) =>
+      module.sendCreateGroup ? module.sendCreateGroup : null
+  },
+  {
+    id: 'sendAddParticipants',
+    conditions: (module) =>
+      module.sendAddParticipants ? module.sendAddParticipants : null
+  },
+  {
+    id: 'sendRemoveParticipants',
+    conditions: (module) =>
+      module.sendRemoveParticipants ? module.sendRemoveParticipants : null
+  },
+  {
+    id: 'sendPromoteParticipants',
+    conditions: (module) =>
+      module.sendPromoteParticipants ? module.sendPromoteParticipants : null
+  },
+  {
+    id: 'sendDemoteParticipants',
+    conditions: (module) =>
+      module.sendDemoteParticipants ? module.sendDemoteParticipants : null
+  },
+  {
+    id: 'checkNumber',
+    conditions: (module) =>
+      module.default &&
+      typeof module.default.toString === 'function' &&
+      module.default.toString().includes('Should not reach queryExists MD')
+        ? module.default
+        : null,
+  },
 ];