+++ src/types/WAPI.d.ts
@@ -81,7 +81,16 @@ interface WAPI {
     contactId: string | string[]
   ) => GroupCreation;
   deleteConversation: (chatId: string) => boolean;
-  deleteMessages: (contactId: string, messageId: string[]) => Promise<object>;
+  deleteMessagesAll: (
+    contactId: string,
+    messageId: string[],
+    onlyLocal: boolean
+  ) => Promise<object>;
+  deleteMessagesMe: (
+    contactId: string,
+    messageId: string[],
+    onlyLocal: boolean
+  ) => Promise<object>;
   demoteParticipant: (groupId: string, contactId: string | string[]) => void;
   downloadFile: (data: string) => Promise<string | boolean>;
   downloadMedia: (messageId: string) => Promise<string>;
@@ -220,6 +229,7 @@ interface WAPI {
     buttonText: string,
     menu: Array<any>
   ) => Promise<Object>;
+  checkChat: (contactId: string) => Promise<Object>;
   sendMessageMentioned: (...args: any) => any;
   sendMessageOptions: (
     chat: any,
@@ -237,6 +247,7 @@ interface WAPI {
     url: string,
     title: string,
     description: string,
+    text: string,
     chatId: string
   ) => void;
   sendMute: (id: string, time: number, type: string) => Promise<object>;